refactor(backup): extract import validation into backup/validate.rs
